requirements.txt
# Core scientific computing
numpy>=1.20.0
pandas
scipy>=1.6.0

# Machine Learning
scikit-learn>=0.24.0
tensorflow
torch

# Neuroimaging tools
nibabel>=3.2.0
nilearn>=0.8.0
antspyx
pyradiomics>=3.0.0

# Visualization
matplotlib
seaborn

# Jupyter environment
jupyter
ipykernel

# Image processing
SimpleITK
opencv-python

# Statistical analysis
statsmodels>=0.12.0
lifelines>=0.26.0

# Utilities
tqdm
networkx>=2.5
pyyaml>=5.4.0 